refactor(core-run): split error into a single file
