+++ core/run/src/error.rs
@@ -1,0 +1,33 @@
+use protocol::{Display, From, ProtocolError, ProtocolErrorKind};
+
+#[derive(Debug, Display, From)]
+pub enum MainError {
+    #[display(fmt = "The axon configuration read failed {:?}", _0)]
+    ConfigParse(common_config_parser::ParseError),
+
+    #[display(fmt = "{:?}", _0)]
+    Io(std::io::Error),
+
+    #[display(fmt = "Toml fails to parse genesis {:?}", _0)]
+    GenesisTomlDe(toml::de::Error),
+
+    #[display(fmt = "crypto error {:?}", _0)]
+    Crypto(common_crypto::Error),
+
+    #[display(fmt = "{:?}", _0)]
+    Utf8(std::string::FromUtf8Error),
+
+    #[display(fmt = "{:?}", _0)]
+    JSONParse(serde_json::error::Error),
+
+    #[display(fmt = "other error {:?}", _0)]
+    Other(String),
+}
+
+impl std::error::Error for MainError {}
+
+impl From<MainError> for ProtocolError {
+    fn from(error: MainError) -> ProtocolError {
+        ProtocolError::new(ProtocolErrorKind::Main, Box::new(error))
+    }
+}
